// src/type.ts
export type ListNode = {
  /**
   * @title type
   */
  type: "bulletList" | "orderedList";

  /**
   * @title content
   */
  content: ListItemNode[];
};

export type ListItemNode = {
  /**
   * @title type
   */
  type: "listItem";

  /**
   * content must contain at least one of the following nodes:
   * - bulletList
   * - codeBlock with no marks
   * - mediaSingle
   * - orderedList
   * - paragraph with no marks
   *
   * @title content
   */
  content: (
    | ListNode
    | IJiraService.CodeBlockNode
    | IJiraService.MediaSingleNode
    | IJiraService.ParagraphContentWithoutNoMarks
  )[];
};

export type LookUp<U extends { type: string }, T> = U extends { type: T } ? U : never;

export namespace IJiraService {
// ...
  export interface ICreateCommentByMarkdownInput extends Omit<ICreateCommentInput, "body"> {
    /**
     * @title body of comment
     */
    body: Omit<Comment["body"], "content"> & {
      /**
       * You must use markdown format string.
       *
       * It is recommended to contain as much detail as possible on the issue raised by the user,
       * so that the next person who reads this issue can see the summary and description of this issue to resolve the issue.
       *
       * @title contents of description
       */
      content: string;
    };
  }
// ...
  export type Text = {
    /**
     * @title text type
     */
    type: "text";

    /**
     * @title content of this text content
     */
    text: string;

    /**
     * It means the emphasis of the markdown format, and it means that there is a string between the backticks.
     *
     * @title marks
     */
    marks?: Mark[];
  };

  /**
   * The blockquote node is a container for quotes.
   * blockquote is a top-level block node.
   *
   * @title Blockquote node
   */
  export type BlockquoteNode = {
    /**
     * @title blockquote type
     */
    type: "blockquote";

    /**
     * content must contain array of one or more of the following nodes:
     * - paragraph with no marks.
     * - bulletList
     * - orderedList
     *
     * @title content
     */
    content: (ParagraphContentWithoutNoMarks | ListNode)[] & tags.MinItems<1>;
  };

  /**
   * @title code block
   */
  export type CodeBlockNode = {
    /**
     * @title type
     */
    type: "codeBlock";

    /**
     * If you do not specify a programming language, this property may not exist.
     *
     * @title attrs
     */
    attrs?: {
      /**
       * @title programming language name
       */
      language?: string;
    };

    /**
     * content takes an array of one or more text nodes without marks.
     *
     * @title code content
     */
    content?: Array<{
      /**
       * @title type
       */
      type: "text";

      /**
       * @title text includeing code
       */
      text: string;
    }> &
      tags.MinItems<1> &
      tags.MaxItems<1>;
  };

  /**
   * @title emoji node
   */
  export type EmojiNode = {
    /**
     * @title emoji type
     */
    type: "emoji";

    /**
     * @title attributes of emoji node
     */
    attrs: {
      /**
       * Emoji service ID of the emoji
       *
       * For custom emojis by atlasian or user, you may have an ID.
       */
      id?: string;

      /**
       * @title icon name
       *
       * A string exists between a colon and a colon, meaning a name for representing the emoji.
       */
      shortName: `:${string}:`;

      /**
       * @title emoji icon
       */
      text?: string;
    };
  };

  /**
   * @title hard break node
   */
  export type HardBreakNode = {
    /**
     * @title hardBreak type
     */
    type: "hardBreak";

    /**
     * @title attributes of hard break node
     */
    attrs?: {
      /**
       * @title text
       *
       * It can be only `\n` text for braking.
       */
      text?: "\n";
    };
  };

  /**
   * @title heading node
   *
   * It means h1, h2, h3, h4, h5, h6 node.
   */
  export type HeadingNode = {
    /**
     * @title heading type
     */
    type: "heading";

    /**
     * @title content
     *
     * Heading node's content can be combined with only inline nodes.
     */
    content: InlineNode[];

    /**
     * @title attributes of heading node
     */
    attrs: {
      /**
       * level represents the depth of the heading following the same convention as HTML: when level is set to 1 it's the equivalent of <h1>.
       */
      level: 1 | 2 | 3 | 4 | 5 | 6;
    };
  };
// ...
  export type InlineCardNode = {
    /**
     * @title inline card type
     */
    type: "inlineCard";

    /**
     * @title attributes of inline card node
     */
    attrs: {
      /**
       * Indicates the address value that the inline card will represent.
       * To allow you to move when you click on the card, you need to put a link in advance.
       *
       * @title url
       */
      url: string & tags.Format<"iri">;

      /**
       * @title representation of the link
       */
      data?: any;
    };
  };
// ...
  export type MediaGroupNode = {
    /**
     * @title type
     */
    type: "mediaGroup";

    /**
     * content must contain one or more media nodes.
     *
     * @title content
     */
    content: MediaNode[];
  };

  /**
   * The mediaSingle node is a container for one media item.
   * This node enables the display of the content in full,
   * in contrast to a mediaGroup that is intended for a list of attachments.
   * A common use case is to display an image,
   * but it can also be used for videos,
   * or other types of content usually renderable by a @atlaskit/media card component.
   */
  export type MediaSingleNode = {
    /**
     * @title mediaSingle type
     */
    type: "mediaSingle";

    /**
     * @title attrs
     */
    attrs: {
      /**
       * layout determines the placement of the node on the page. wrap-left and wrap-right provide an image floated to the left or right of the page respectively, with text wrapped around it. center center aligns the image as a block, while wide does the same but bleeds into the margins. full-width makes the image stretch from edge to edge of the page.
       */
      layout: "wrap-left" | "center" | "wrap-right" | "wide" | "full-width" | "align-start" | "align-end";

      /**
       * width determines the width of the image as a percentage of the width of the text content area. Has no effect if layout mode is wide or full-width.
       *
       * @title width
       */
      width?: number & tags.Minimum<0> & tags.Maximum<100>;

      /**
       * widthType [optional] determines what the "unit" of the width attribute is presenting. Possible values are pixel and percentage. If the widthType attribute is undefined, it fallbacks to percentage.
       *
       * @title widthType
       */
      widthType?: ("pixel" | "percentage") & tags.Default<"percentage">;
    };

    /**
     * @title media
     *
     * only single of media node type
     */
    content: Array<MediaNode> & tags.MinItems<1> & tags.MaxItems<1>;
  };

  /**
   * @title content with mention
   */
  export type MentionNode = {
    /**
     * @title mention type
     */
    type: "mention";

    // text?: never;

    /**
     * @title content of this mention content
     */
    attrs: {
      /**
       * @title id
       *
       * add any string like as uuid
       */
      id?: string;

      /**
       * @title Who is mentioned
       *
       * It means a string that connects @ and the user's name
       */
      text: `@${string}`;

      /**
       * @title accessLevel
       */
      accessLevel?: string;
    };
  };

  /**
   * The panel node is a container that highlights content.
   */
  export type PanelNode = {
    /**
     * @title type
     */
    type: "panel";

    /**
     * @title attrs
     */
    attrs: {
      /**
       * @title panelType
       */
      panelType: "info" | "note" | "warning" | "success" | "error";
    };

    /**
     * @title content
     */
    content: (ListNode | HeadingNodeWithoutMarks | ParagraphContentWithoutNoMarks)[];
  };

  /**
   * @title paragraph type
   * The paragraph node is a container for a block of formatted text delineated by a carriage return. It's the equivalent of the HTML <p> tag.
   */
  export type ParagraphNode = {
    /**
     * @title paragraph type
     */
    type: "paragraph";

    /**
     * @title attrs
     */
    attrs?: never | Record<string, never>;

    /**
     * @title content
     *
     * If you want to make a new line, there will be an empty array.
     */
    content: InlineNode[];
  };

  /**
   * @title paragraph type
   */
  export type ParagraphContentWithoutNoMarks = {
    /**
     * @title paragraph type
     */
    type: "paragraph";

    /**
     * @title attrs
     */
    attrs?: never | Record<string, never>;

    /**
     * @title content
     *
     * If you want to make a new line, there will be an empty array.
     */
    content: (EmojiNode | HardBreakNode | InlineCardNode | MentionNode | Omit<Text, "marks">)[];
  };

  /**
   * The rule node represents a divider, it is equivalent to the HTML <hr/> tag.
   */
  export type RuleNode = {
    /**
     * @title type
     */
    type: "rule";
  };

  /**
   * The table node provides a container for the nodes that define a table.
   *
   * Note: only supported on web and desktop. Mobile rendering support for tables is not available.
   */
  export type TableNode = {
    /**
     * @title type
     */
    type: "table";

    /**
     * @title content
     */
    content: TableRowNode[];

    /**
     * @title attrs
     */
    attrs?: {
      /**
       * When isNumberColumnEnabled is set to 'true' the first table column provides numbering for the table rows.
       *
       * @title isNumberColumnEnabled
       */
      isNumberColumnEnabled?: (true | false) & tags.Default<true>;

      /**
       * width sets the length (in pixels) of the table on the page.
       * This value is independent of the table's column width, this allows control of the table's overflow.
       * It supersedes the existing layout attribute and will be used instead of it at runtime.
       * If width is not provided the editor will convert layout to pixels (default=760, wide=960 and full-width=1800).
       * Although no minimum and maximum width is enforced it is recommended to follow these guidelines:
       *
       * - Minimum width
       *
       * - 1 column table = 48px
       * - 2 column table = 96px
       * - 3 column table = 144px
       * - > 3 column table = 144px
       * - Maximum width
       * - 1800
       *
       * @title width
       */
      width?: number & tags.Minimum<0>;

      /**
       * layout determines the alignment of a table in the full page editor, relevant to the line length. Currently only center and left alignment options are supported.
       * The layout values are mapped as follows:
       * - 'center' : will align the table to the center of page, its width can be larger than the line length
       * - 'align-start' : will align the table left of the line length, its width cannot be larger than the line length
       *
       * @title layout
       */
      layout?: "center" | "align-start";

      /**
       * displayMode attribute controls how tables adapt to narrow screens:
       * When displayMode is set to 'default' or left unset, the table's columns will automatically scale down to accommodate narrow screens, with a maximum reduction of up to 40%.
       * When displayMode is set to 'fixed', the table's columns will maintain their original width, regardless of screen size.
       *
       * @title displayMode
       */
      displayMode?: "default" | "fixed";
    };
  };

  export type TableCellNode = {
    /**
     * @title type
     */
    type: "tableCell";

    /**
     * @title content
     */
    content: (
      | BlockquoteNode
      | ListNode
      | CodeBlockNode
      | HeadingNode
      | MediaGroupNode
      | PanelNode
      | ParagraphNode
      | RuleNode
    )[];

    /**
     * @title attrs
     */
    attrs?: {
      /**
       * Short or long hex color code or HTML color name
       *
       * @title background
       */
      background?: string;

      /**
       * colspan defines the number of columns the cell spans.
       *
       * @title colspan
       */
      colspan?: number & tags.Type<"uint64"> & tags.Default<1>;

      /**
       * defines the width of the column or,
       * where the cell spans columns, the width of the columns it spans in pixels.
       * The length of the array should be equal to the number of spanned columns.
       * 0 is permitted as an array value if the column size is not fixed,
       * for example, a cell merged across 3 columns where one unfixed column is surrounded by two fixed might be represented as `[120, 0, 120].
       *
       * @title colwidth
       */
      colwidth?: Array<number & tags.Type<"uint64">> & tags.MinItems<3> & tags.MaxItems<3>;

      /**
       * rowspan defines the number of rows a cell spans.
       *
       * @title rowspan
       */
      rowspan?: number & tags.Type<"uint64"> & tags.Default<1>;
    };
  };
// ...
  export type TableRowNode = {
    /**
     * @title type
     */
    type: "tableRow";

    /**
     * content takes an array of one or more tableHeader or tableCell nodes.
     *
     * @title content
     */
    content: (TableHeaderNode | TableCellNode)[];
  };
// ...
  export type TopLevelBlockNode =
    | BlockquoteNode
    | ListNode
    | CodeBlockNode
    | HeadingNode
    | MediaGroupNode
    | MediaSingleNode
    | PanelNode
    | ParagraphNode
    | RuleNode
    | TableNode;

  /**
   * An inline node is the nodes that belong to Paragrap and are used to write.
   * Typically used as a child node for a Paragrap node.
   *
   * @title inlineNode
   */
  export type InlineNode = EmojiNode | HardBreakNode | InlineCardNode | MentionNode | Text;
// ...
}